+++ src/zaakmagazijn/api/tests/test_geef_besluit_details.py
@@ -1,4 +1,3 @@
-
 from lxml import etree
 from zeep.xsd.const import Nil