+++ src/zaakmagazijn/auditlog_extension/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'zaakmagazijn.auditlog_extension.apps.AuditLogExtConfig'